pyproject.toml
[project]
name = "email-management-tool"
version = "1.0.0"
description = "Email Management Tool - Enterprise email moderation system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "aiosmtpd>=1.4.0",
    "flask-login>=0.6.0",
    "flask-wtf>=1.2.0",
    "werkzeug>=3.0.0",
    "bcrypt>=4.0.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "python-dateutil>=2.8.0",
    "jinja2>=3.1.0",
    "markupsafe>=2.1.0",
    "itsdangerous>=2.1.0",
    "click>=8.1.0",
    "colorama>=0.4.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-flask>=1.2.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
    "pylint>=2.17.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]
test = [
    "faker>=19.0.0",
    "factory-boy>=3.3.0",
    "responses>=0.23.0",
    "freezegun>=1.2.0",
    "coverage>=7.2.0",
    "pytest-html>=3.2.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "locust>=2.15.0",
    "selenium>=4.10.0",
    "pytest-bdd>=6.1.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.23.0",
    "myst-parser>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--maxfail=1",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
    "smtp: SMTP-related tests",
    "imap: IMAP-related tests",
    "security: Security tests",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/test_*.py", "*/__pycache__/*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[tool.pylint.messages_control]
disable = "C0111,R0903,R0913,W0613"
